+++ build.rs
@@ -1,5 +1,11 @@
-use clap_complete::{generate_to, shells::Bash};
+use clap_complete::{
+    generate_to,
+    shells::{Bash, Elvish, Fish, PowerShell, Zsh},
+    Generator,
+};
+use std::path::PathBuf;
 use std::{env, fs, path};
+use std::{ffi::OsString, fmt::format};
 
 use std::io::Error;
 
@@ -13,13 +19,13 @@ fn main() -> Result<(), Error> {
 
     let mut cmd = build_cli();
     //this is only generated when the git ref changes???
-    let path = generate_to(
-        Bash, &mut cmd, // We need to specify what generator to use
-        "artem",  // We need to specify the bin name manually
-        &out_dir, // We need to specify where to write to
-    )?;
+    let mut shells_paths = Vec::with_capacity(5);
 
-    println!("cargo:warning=completion file is generated: {:?}", path);
+    shells_paths.push(generate_shell_completion(&mut cmd, &out_dir, Bash));
+    shells_paths.push(generate_shell_completion(&mut cmd, &out_dir, PowerShell));
+    shells_paths.push(generate_shell_completion(&mut cmd, &out_dir, Zsh));
+    shells_paths.push(generate_shell_completion(&mut cmd, &out_dir, Fish));
+    shells_paths.push(generate_shell_completion(&mut cmd, &out_dir, Elvish));
 
     //get output file location
     let project_dir = match env::var_os("CARGO_MANIFEST_DIR") {
@@ -31,14 +37,27 @@ fn main() -> Result<(), Error> {
     println!("cargo:warning=creating deployment directory");
     fs::create_dir_all(project_dir.join("deployment/deb")).expect("Failed to create project dir");
 
-    println!(
-        "cargo:warning=copying completion file to: {:?}",
-        project_dir.join("/deployment/deb/artem.bash")
-    );
+    for path in shells_paths {
+        if path.is_err() {
+            continue;
+        }
 
-    //copy generated completion script to deployment dir
-    fs::copy(&path, project_dir.join("deployment/deb/artem.bash"))
-        .expect("failed to copy completion script");
+        let unwrapped = &path.unwrap();
+
+        let extension = match &unwrapped.extension() {
+            Some(v) => v.to_str().unwrap(),
+            None => "",
+        };
+
+        let completion_path = project_dir.join(format!("/deployment/deb/artem.{}", extension));
+        println!(
+            "cargo:warning=copying completion file to: {}",
+            completion_path.to_str().unwrap()
+        );
+
+        //copy generated completion script to deployment dir
+        fs::copy(&unwrapped, completion_path).expect_err("failed to copy completion script");
+    }
 
     let man = clap_mangen::Man::new(cmd);
     let mut buffer: Vec<u8> = Default::default();
@@ -59,4 +78,22 @@ fn main() -> Result<(), Error> {
         .expect("failed to copy man page");
 
     Ok(())
+}
+
+fn generate_shell_completion<T>(
+    cmd: &mut Command,
+    out_dir: &OsString,
+    shell: T,
+) -> Result<PathBuf, Error>
+where
+    T: Generator,
+{
+    //generate shell completions
+    let path = generate_to(
+        shell, cmd,      // We need to specify what generator to use
+        "artem",  // We need to specify the bin name manually
+        &out_dir, // We need to specify where to write to
+    )?;
+    println!("cargo:warning=completion file is generated: {:?}", &path);
+    Ok(path)
 }